fix: handle conflicts by rename type
